Replace p. with page. and p.page. with page.
This part of the documentation was confusing since a `p` variable wasn't mentioned previously. After using this library I realized what it was really talking about is the object in the `ph.createPage` callback, previously referred to as `page`. Therefore, I thought it made the most sense to refer to it as `page` here as well.